Move all stats to level up events (+ other events)
