--- src/main/java/mc/rpgstats/component/FarmingComponent.java
@@ -1,76 +1,0 @@
-package mc.rpgstats.component;
-
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.text.LiteralText;
-
-public class FarmingComponent implements IStatComponent {
-    private final PlayerEntity player;
-    private int xp = 0;
-    private int level = 0;
-    
-    public FarmingComponent(PlayerEntity player) {
-        this.player = player;
-    }
-    
-    @Override
-    public void readFromNbt(CompoundTag tag) {
-        this.level = tag.getInt("level");
-        this.xp = tag.getInt("xp");
-    }
-    
-    @Override
-    public void writeToNbt(CompoundTag tag) {
-        tag.putInt("xp", this.xp);
-        tag.putInt("level", this.level);
-    }
-    
-    @Override
-    public Entity getEntity() {
-        return player;
-    }
-    
-    @Override
-    public int getXP() {
-        return this.xp;
-    }
-    
-    @Override
-    public void setXP(int newXP) {
-        this.xp = newXP;
-    }
-    
-    @Override
-    public int getLevel() {
-        return this.level;
-    }
-    
-    @Override
-    public void setLevel(int newLevel) {
-        this.level = newLevel;
-    }
-    
-    @Override
-    public String getName() {
-        return "farming";
-    }
-    
-    @Override
-    public String getCapName() {
-        return "Farming";
-    }
-    
-    @Override
-    public void onLevelUp(boolean beQuiet) {
-        if (!beQuiet) {
-            player.sendMessage(new LiteralText("§a+1§r Bonemeal efficiency"), false);
-            
-            if (level == 25) {
-                player.sendMessage(new LiteralText("§aNurturing§r - Shift rapidly to grow nearby crops (while holding hoe)"), false);
-            } else if (level == 50) {
-                player.sendMessage(new LiteralText("§aNurturing II§r - Nurturing has increased range"), false);
-            }
-        }
-    }
-}
