+++ src/main/java/mc/rpgstats/main/RPGStatsClient.java
@@ -2,7 +2,6 @@ package mc.rpgstats.main;
 
 import mc.rpgstats.client.screen.RPGStatDisplayGUI;
 import mc.rpgstats.client.screen.RPGStatDisplayScreen;
-import mc.rpgstats.component.IStatComponent;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
@@ -14,7 +13,6 @@ import net.minecraft.util.Pair;
 import org.lwjgl.glfw.GLFW;
 
 import java.util.HashMap;
-import java.util.Optional;
 
 public class RPGStatsClient implements ClientModInitializer {
     public static HashMap<Identifier, String> nameMap = new HashMap<>();
